+++ src/components/App/App.component.jsx
@@ -1,57 +1,32 @@
-import React, { useLayoutEffect } from 'react';
+import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
-
+import { CssBaseline } from '@material-ui/core';
 import AuthProvider from '../../providers/Auth';
-import HomePage from '../../pages/Home';
-import LoginPage from '../../pages/Login';
-import NotFound from '../../pages/NotFound';
-import SecretPage from '../../pages/Secret';
-import Private from '../Private';
-import Fortune from '../Fortune';
-import Layout from '../Layout';
-import { random } from '../../utils/fns';
+import Home from '../../pages/Home';
+import Protected from '../Protected';
+import Video from '../../pages/Video';
+import MainAppBar from '../MainAppBar';
+import LeftDrawer from '../LeftDrawer';
+import useStyles from './AppStyles';
 
 function App() {
-  useLayoutEffect(() => {
-    const { body } = document;
-
-    function rotateBackground() {
-      const xPercent = random(100);
-      const yPercent = random(100);
-      body.style.setProperty('--bg-position', `${xPercent}% ${yPercent}%`);
-    }
-
-    const intervalId = setInterval(rotateBackground, 3000);
-    body.addEventListener('click', rotateBackground);
-
-    return () => {
-      clearInterval(intervalId);
-      body.removeEventListener('click', rotateBackground);
-    };
-  }, []);
+  const classes = useStyles();
 
   return (
-    <BrowserRouter>
-      <AuthProvider>
-        <Layout>
+    <AuthProvider>
+      <BrowserRouter>
+        <div className={classes.root}>
+          <CssBaseline />
+          <MainAppBar />
+          <LeftDrawer />
           <Switch>
-            <Route exact path="/">
-              <HomePage />
-            </Route>
-            <Route exact path="/login">
-              <LoginPage />
-            </Route>
-            <Private exact path="/secret">
-              <SecretPage />
-            </Private>
-            <Route path="*">
-              <NotFound />
-            </Route>
+            <Route exact path="/" component={Home} />
+            <Route path="/player/:id" component={Video} />
+            <Protected path="/favorites" component={Home} />
           </Switch>
-          <Fortune />
-        </Layout>
-      </AuthProvider>
-    </BrowserRouter>
+        </div>
+      </BrowserRouter>
+    </AuthProvider>
   );
 }
