+++ src/components/App/AppStyles.js
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  root: {
+    display: 'flex',
+    flexGrow: 1,
+  },
+}));
+
+export default useStyles;
