+++ src/components/Protected/Protected.component.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Redirect, Route } from 'react-router-dom';
+
+import { useAuth } from '../../providers/Auth';
+
+function Protected(props) {
+  const { authenticated } = useAuth();
+
+  return authenticated ? <Route {...props} /> : <Redirect to="/" />;
+}
+
+export default Protected;
