--- src/pages/Main/index.js
@@ -1,1 +1,0 @@
-export { default } from './Main.page';
